+++ supertokens_python/recipe/jwt/asyncio/__init__.py
@@ -13,12 +13,13 @@
 # under the License.
 from typing import Any, Dict, Union
 
-from supertokens_python.recipe.jwt.interfaces import (CreateJwtResult,
+from supertokens_python.recipe.jwt.interfaces import (CreateJwtResultOk,
+                                                      CreateJwtResultUnsupportedAlgorithm,
                                                       GetJWKSResult)
 from supertokens_python.recipe.jwt.recipe import JWTRecipe
 
 
-async def create_jwt(payload: Union[None, Dict[str, Any]] = None, validity_seconds: Union[None, int] = None, user_context: Union[Dict[str, Any], None] = None) -> CreateJwtResult:
+async def create_jwt(payload: Union[None, Dict[str, Any]] = None, validity_seconds: Union[None, int] = None, user_context: Union[Dict[str, Any], None] = None) -> Union[CreateJwtResultOk, CreateJwtResultUnsupportedAlgorithm]:
     if user_context is None:
         user_context = {}
     if payload is None: